% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{runExample}
\usage{
runExample(example = "02-plugins")
}
\arguments{
\item{example}{}
}
\description{
Run example Shiny App
}
\examples{
\dontrun{
  shinyPlugins::runExample("02-plugins")
}
}
\author{
Kamil Foltyński
}
