% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plugins.R
\docType{class}
\name{shiny_plugins}
\alias{shiny_plugins}
\title{shiny_plugins}
\format{\code{\link{R6Class}} object}
\usage{
shiny_plugins
}
\arguments{
\item{ui.pattern}{prefix to find **UI** functions of modules}

\item{server.pattern}{prefix to find **server** functions of modules}
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class providing object of \code{\link{R6Class}} to easily and automatically implement Shiny plugins (modules)
}
\section{Methods}{

\describe{
  \item{\code{getPluginsList()}}{Get list of found plugins}
  \item{\code{getTabItems()}}{Get list of \code{\link[shinydashboard]{tabItem}} with content of plugins}
  \item{\code{getMenuItem()}}{Get \code{\link[shinydashboard]{menuItem}} containing plugins}

  \item{\code{getModulesUI()}}{Get names of modules **UI** functions}
  \item{\code{getModulesServer()}}{Get names of modules **server** functions}

  }
}

\examples{
\dontrun{
  library(shinyPlugins)
  plugins <- shiny_plugins$new(where = ls("package:shinyPlugins"))
}

}
\keyword{plugins}
