% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{shinyPluginUI_example}
\alias{shinyPluginUI_example}
\alias{shinyPluginServer_example}
\title{shinyPlugin_exampleUI}
\usage{
shinyPluginUI_example(id)

shinyPluginServer_example(input, output, session)
}
\arguments{
\item{id}{}

\item{session}{}
}
\description{
shinyPlugin_exampleUI

shinyPlugin_example
}
