% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{shinyPlugin_example2UI}
\alias{shinyPlugin_example2UI}
\alias{shinyPlugin_example2}
\title{shinyPlugin_example2UI}
\usage{
shinyPlugin_example2UI(id)

shinyPlugin_example2(input, output, session)
}
\arguments{
\item{id}{}

\item{session}{}
}
\description{
shinyPlugin_example2UI

shinyPlugin_example2
}
