% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plugins.R
\name{shinyPlugin_exampleUI}
\alias{shinyPlugin_exampleUI}
\alias{shinyPlugin_example}
\title{shinyPlugin_exampleUI}
\usage{
shinyPlugin_exampleUI(id)

shinyPlugin_example(input, output, session)
}
\arguments{
\item{id}{}

\item{session}{}
}
\description{
shinyPlugin_exampleUI

shinyPlugin_example
}
