% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{shinyPluginUI_example2}
\alias{shinyPluginUI_example2}
\alias{shinyPluginServer_example2}
\title{shinyPlugin_example2UI}
\usage{
shinyPluginUI_example2(id)

shinyPluginServer_example2(input, output, session)
}
\arguments{
\item{id}{}

\item{session}{}
}
\description{
shinyPlugin_example2UI

shinyPlugin_example2
}
